requirements.txt
langchain
langgraph
langchain_ollama
langchain-community
langdetect
google-search-results
python-dotenv